Reintroduced zooming
Added the previous zooming functionality to the new scrollevent
handler. Zooming can now be done by pushing ctrl while scrolling.
Scrolling without pushing ctrl will scroll the graph horizontally.
