--- src/main/java/gui/PannableCanvas.java
@@ -1,87 +1,0 @@
-package gui;
-
-import javafx.beans.property.DoubleProperty;
-import javafx.beans.property.SimpleDoubleProperty;
-import javafx.scene.canvas.Canvas;
-import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.layout.Pane;
-import javafx.scene.paint.Color;
-
-/**
- * LIKELY TO BE REPLACED WITH A SCROLLPANE.
- * 
- * A custom Pane which can be panned over.
- */
-@SuppressWarnings("restriction")
-public class PannableCanvas extends Pane {
-
-    DoubleProperty myScale = new SimpleDoubleProperty(1.0);
-
-    public PannableCanvas() {
-        setPrefSize(600, 600);
-
-        // add scale transform
-        scaleXProperty().bind(myScale);
-        scaleYProperty().bind(myScale);
-    }
-    
-    public PannableCanvas(int hsize, int vsize) {
-    	this.setPrefSize(hsize, vsize);
-    	this.setStyle("-fx-border-color: blue;");
-    	
-        // add scale transform
-        this.scaleXProperty().bind(myScale);
-        this.scaleYProperty().bind(myScale);
-    }
-
-    /**
-     * Add a grid to the canvas, send it to back
-     */
-    public void addGrid() {
-
-        double w = getBoundsInLocal().getWidth();
-        double h = getBoundsInLocal().getHeight();
-
-        // add grid
-        Canvas grid = new Canvas(w, h);
-
-        // don't catch mouse events
-        grid.setMouseTransparent(true);
-
-        GraphicsContext gc = grid.getGraphicsContext2D();
-
-        gc.setStroke(Color.GRAY);
-        gc.setLineWidth(1);
-
-        // draw grid lines
-        double offset = 50;
-        for (double i = offset; i < w; i += offset) {
-            gc.strokeLine( i, 0, i, h);
-            gc.strokeLine( 0, i, w, i);
-        }
-
-        getChildren().add( grid);
-
-        grid.toBack();
-    }
-
-    public double getScale() {
-        return myScale.get();
-    }
-
-    public void setScale( double scale) {
-        myScale.set(scale);
-    }
-
-    public void reset() { 
-    	PannableCanvas preset = new PannableCanvas();
-    	this.setScale(preset.getScale());
-    	this.setTranslateX(100);
-    	this.setTranslateY(100);
-    }
-    
-    public void setPivot(double xc, double yc) {
-        setTranslateX(getTranslateX() - xc);
-        setTranslateY(getTranslateY() - yc);
-    }
-}
